+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 #### Next Release
 - Should highlight selected verse when opening from notes
+- Fixed crash when downloading translation
 
 #### v0.3.0 (2019-05-17)
 - Added event tracking